requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.0.12
decorator==5.1.1
Django==4.0.5
djangorestframework==3.13.1
future==0.18.2
gunicorn==20.1.0
idna==3.3
iso8601==1.0.2
owner==0.0.10
psycopg2-binary==2.9.3
PyJWT==2.4.0
pytz==2022.1
PyYAML==6.0
requests==2.28.0
self==2020.12.3
serial==0.0.97
sqlparse==0.4.2
urllib3==1.26.9
